refactor: 회원가입 및 프로필 설저에서 사용하는 nickname 컴포넌트 common으로 이동
